[refactor] #142 auth Controller 리팩토링
- checktoken 미들웨어로 빼줌
- service로 넘겨줄때 의미 알 수 있게 변수에 한번 넣어서 넣어줌
- 절대경로 변경
- response 만들어서 반환
