+++ BE/src/controllers/auth.ts
@@ -1,33 +1,25 @@
-import createError from 'http-errors';
-import { Context, Next } from 'koa';
+import { Context } from 'koa';
 
-import serviceAuthCheck from '@services/auth/check';
-import serviceLogin from '@services/auth/login';
+import serviceLogin from '@/services/auth/login';
+
+import response from '@/utils/response';
 
 const login = async (ctx: Context): Promise<Context['body']> => {
-  const tokens = await serviceLogin.login(ctx.request.body);
-  ctx.body = tokens;
+  const loginInfo = ctx.request.body;
+  const tokens = await serviceLogin.login(loginInfo);
+  const res = response(200, { ...tokens });
+  ctx.status = res.status;
+  ctx.body = res.data;
   return ctx.body;
 };
 
 const refresh = async (ctx: Context): Promise<Context['body']> => {
-  const newAccessToken = await serviceLogin.refresh(ctx.request.body);
-  ctx.body = { newAccessToken };
+  const requestInfo = ctx.request.body;
+  const accessToken = await serviceLogin.refresh(requestInfo);
+  const res = response(200, { accessToken });
+  ctx.status = res.status;
+  ctx.body = res.data;
   return ctx.body;
 };
 
-const checkToken = async (
-  ctx: Context,
-  next: Next,
-): Promise<Context['body']> => {
-  const user = await serviceAuthCheck.checkToken(ctx.header);
-  if (!user) {
-    const jwtError = createError(401, 'unauthorized');
-    throw jwtError;
-  }
-  ctx.accessToken = serviceLogin.slidingSession(user);
-  ctx.user = user;
-  await next();
-};
-
-export default { login, refresh, checkToken };
+export default { login, refresh };
