+++ BE/src/routes/api/index.ts
@@ -1,6 +1,7 @@
 import Router from 'koa-router';
 
-import authController from '@/controllers/auth';
+import authMiddleware from '@/middlewares/check-token';
+
 import authRouter from '@/routes/api/auth';
 import accountBookRouter from '@/routes/api/account-book';
 import defaultPaymentMethodRouter from '@/routes/api/defaultPaymentMethod';
@@ -11,7 +12,7 @@ const router = new Router();
 // api/
 
 router.use('/auth', authRouter.routes());
-router.use(authController.checkToken);
+router.use(authMiddleware.checkToken);
 router.use('/accountbook', accountBookRouter.routes());
 router.use('/social/accountbook', socialRouter.routes());
 router.use('/defaultPaymentMethod', defaultPaymentMethodRouter.routes());
